+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This Changelog is automatically stamped by `cut_release.sh` as part of creating 
 - An auth router for tasks related to authentication.
 - Endpoint to retrieve the n most relevant snippets for a user query.
 - Bootstrap Icons 1.11.3
+- SCSS compilation pipeline to customise Bootstrap's palette.
 
 ###  Removed